add CommonRequestQueryFilterResolver and InspectMappers for enhanced request handling
